+++ addons/logistix/models/__init__.py
@@ -1,2 +1,14 @@
-# LogistiX Models
-# Imports will be added as models are developed.
+# 1. Base Mixins
+from . import mixins
+
+# 2. Native Model Extensions
+from . import partner
+from . import employee
+from . import vehicle
+
+# 3. Custom Models
+from . import trip
+from . import vehicle_document
+from . import driver_document
+from . import notification
+from . import audit_log