+++ addons/logistix/models/mixins/__init__.py
@@ -1,0 +1,3 @@
+from . import document_mixin
+from . import timestamp_mixin
+from . import audit_mixin